+++ lab4/src/find_min_max.h
@@ -1,8 +1,9 @@
 #ifndef FIND_MIN_MAX_H
 #define FIND_MIN_MAX_H
+#include <stddef.h>
 
 #include "utils.h"
 
-struct MinMax GetMinMax(int *array, unsigned int begin, unsigned int end);
+struct MinMax GetMinMax(int *array, size_t begin, size_t end);
 
 #endif
